fix: Wait for green light after missing stop line waypoint.
Previous feature that would enable late braking even if the car was going to
miss the stop line by certain margin did not entirely solve the red light
running problem. Car would ignore the light once it passes the stop target
waypoint causing it to stop and go instead of waiting for green light in late
braking situation. Making the car observe the light even if it is within 2
waypoints beyond it's corresponding stop target waypoint.